fix bug in simple prof(singulaton)
