+++ common/CMakeLists.txt
@@ -1,2 +1,3 @@
 file(GLOB srcs "*.cc")
-add_library(${PROJECT_NAME}_common STATIC ${srcs} ./performance/simple_prof.cc)
+file(GLOB perf_srcs "performance/*.cc")
+add_library(${PROJECT_NAME}_common STATIC ${srcs} ${perf_srcs})
